refactor: move feature configuration components to a separate package for better modularity
