+++ packages/webtrit_appearance_theme/lib/converters/bottom_menu_tab_type_converter.dart
@@ -1,6 +1,7 @@
 import 'package:json_annotation/json_annotation.dart';
 
-import '../models/models.dart';
+import 'package:webtrit_appearance_theme/models/models.dart';
+
 
 class BottomMenuTabTypeConverter implements JsonConverter<BottomMenuTabType, String> {
   const BottomMenuTabTypeConverter();
