+++ packages/webtrit_appearance_theme/lib/models/features_config/features_config.dart
@@ -1,0 +1,4 @@
+export 'app_config.dart';
+export 'bottom_menu_tab_type.dart';
+export 'elevated_button_style_type.dart';
+export 'metadata.dart';
