+++ packages/webtrit_appearance_theme/lib/models/models.dart
@@ -1,9 +1,6 @@
-export 'app_config.dart';
-export 'bottom_menu_tab_type.dart';
 export 'color_scheme.config.dart';
 export 'custom_color.dart';
-export 'elevated_button_style_type.dart';
-export 'metadata.dart';
+export 'features_config/features_config.dart';
 export 'theme_page_config.dart';
 export 'theme_settings.dart';
 export 'theme_widget_config.dart';
