+++ packages/webtrit_appearance_theme/lib/models/theme_settings.dart
@@ -1,6 +1,5 @@
 import 'package:freezed_annotation/freezed_annotation.dart';
 
-import 'app_config.dart';
 import 'color_scheme.config.dart';
 import 'theme_page_config.dart';
 import 'theme_widget_config.dart';
@@ -21,8 +20,6 @@ class ThemeSettings with _$ThemeSettings {
     // Pages config
     @Default(ThemePageConfig()) ThemePageConfig themePageLightConfig,
     @Default(ThemePageConfig()) ThemePageConfig themePageDarkConfig,
-    // Feature access
-    AppConfig? appConfig,
   }) = _ThemeSettings;
 
   factory ThemeSettings.fromJson(Map<String, dynamic> json) => _$ThemeSettingsFromJson(json);
